dashboard/backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.4.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
meraki==1.43.0
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
tzdata==2024.1
urllib3==1.26.18
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
